[fix]诊疗以及预约列表增加分页跟排序

+++ medical-provider/src/main/java/com/xczhihui/medical/doctor/service/impl/RemoteTreatmentServiceImpl.java
@@ -1,5 +1,14 @@
 package com.xczhihui.medical.doctor.service.impl;
 
+import static com.xczhihui.common.util.enums.TreatmentInfoApplyStatus.APPLY_PASSED;
+
+import java.text.ParseException;
+import java.text.SimpleDateFormat;
+import java.util.*;
+
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.stereotype.Service;
 
 import com.baomidou.mybatisplus.plugins.Page;
 import com.jayway.jsonpath.JsonPath;
@@ -23,16 +32,6 @@ import com.xczhihui.medical.doctor.vo.MedicalDoctorVO;
 import com.xczhihui.medical.doctor.vo.TreatmentVO;
 import com.xczhihui.medical.enrol.mapper.MedicalEntryInformationMapper;
 import com.xczhihui.medical.exception.MedicalException;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.stereotype.Service;
-
-import java.text.ParseException;
-import java.text.SimpleDateFormat;
-import java.util.*;
-import java.util.stream.Collectors;
-
-import static com.xczhihui.common.util.enums.TreatmentInfoApplyStatus.APPLY_PASSED;
 
 /**
  * @author hejiwei
@@ -390,67 +389,52 @@ public class RemoteTreatmentServiceImpl implements IRemoteTreatmentService {
     }
 
     @Override
-    public List<TreatmentVO> listByDoctorId(String doctorId) {
+    public List<TreatmentVO> listByDoctorId(String doctorId, int page, int size) {
         List<TreatmentVO> results = new ArrayList<>();
-        List<TreatmentVO> expiredAndFinishedTreatments = remoteTreatmentMapper.selectExpiredAndFinishedByDoctorId(doctorId);
-        List<TreatmentVO> unExpiredAndUnFinishedTreatments = remoteTreatmentMapper.selectUnExpiredAndUnFinishedByDoctorId(doctorId);
 
-        results.addAll(unExpiredAndUnFinishedTreatments.stream()
-                .filter(treatmentVO -> treatmentVO.getStatus() == AppointmentStatus.WAIT_START.getVal() && isCanStartLive(getTreatmentTime(treatmentVO.getDate(), treatmentVO.getStartTime())))
-                .collect(Collectors.toList()));
-        results.addAll(unExpiredAndUnFinishedTreatments.stream()
-                .filter(treatmentVO -> treatmentVO.getStatus() == AppointmentStatus.STARTED.getVal())
-                .collect(Collectors.toList()));
-        results.addAll((unExpiredAndUnFinishedTreatments.stream()
-                .filter(treatmentVO -> treatmentVO.getStatus() == AppointmentStatus.WAIT_START.getVal() && !isCanStartLive(getTreatmentTime(treatmentVO.getDate(), treatmentVO.getStartTime())))
-                .collect(Collectors.toList())));
-        results.addAll(unExpiredAndUnFinishedTreatments.stream()
-                .filter(treatmentVO -> treatmentVO.getStatus() == AppointmentStatus.WAIT_APPLY.getVal())
-                .collect(Collectors.toList()));
+        List<TreatmentVO> treatmentVOS = remoteTreatmentMapper.selectTreatmentByDoctorId(doctorId, new Page<>(page, size));
 
-        results.addAll(expiredAndFinishedTreatments.stream()
-                .filter(treatmentVO -> treatmentVO.getStatus() == AppointmentStatus.EXPIRED.getVal())
-                .collect(Collectors.toList()));
-        results.addAll(expiredAndFinishedTreatments.stream()
-                .filter(treatmentVO -> treatmentVO.getStatus() == AppointmentStatus.FINISHED.getVal())
-                .collect(Collectors.toList()));
+        Iterator<TreatmentVO> iterator = treatmentVOS.iterator();
+        while (iterator.hasNext()) {
+            TreatmentVO treatmentVO = iterator.next();
+            if (treatmentVO.getStatus() == AppointmentStatus.WAIT_START.getVal() && isCanStartLive(getTreatmentTime(treatmentVO.getDate(), treatmentVO.getStartTime()))) {
+                treatmentVO.setStart(true);
+                results.add(treatmentVO);
+                iterator.remove();
+            }
+        }
+        results.addAll(treatmentVOS);
         results.forEach(treatmentVO -> {
             treatmentVO.setTreatmentTime(getTreatmentTime(treatmentVO.getDate(), treatmentVO.getStartTime()));
             handleDate(treatmentVO);
         });
-        cacheService.delete(RedisCacheKey.DOCTOR_TREATMENT_STATUS_CNT_KEY + doctorId);
+        if (page == 1) {
+            cacheService.delete(RedisCacheKey.DOCTOR_TREATMENT_STATUS_CNT_KEY + doctorId);
+        }
         return results;
     }
 
     @Override
-    public List<TreatmentVO> listByUserId(String userId) {
+    public List<TreatmentVO> listByUserId(String userId, int page, int size) {
         List<TreatmentVO> results = new ArrayList<>();
-        List<TreatmentVO> unExpiredAndUnFinishedAppointmentInfoVOS = remoteTreatmentMapper.selectUnExpiredAndUnFinishedByUserId(userId);
-        results.addAll(unExpiredAndUnFinishedAppointmentInfoVOS.stream()
-                .filter(treatmentVO -> (treatmentVO.getStatus() == APPLY_PASSED.getVal() && isCanStartLive(getTreatmentTime(treatmentVO.getDate(), treatmentVO.getStartTime()))))
-                .collect(Collectors.toList()));
-        results.addAll(unExpiredAndUnFinishedAppointmentInfoVOS.stream()
-                .filter(treatmentVO -> (treatmentVO.getStatus() == APPLY_PASSED.getVal() && !isCanStartLive(getTreatmentTime(treatmentVO.getDate(), treatmentVO.getStartTime()))))
-                .collect(Collectors.toList()));
-        results.addAll(unExpiredAndUnFinishedAppointmentInfoVOS.stream()
-                .filter(treatmentVO -> treatmentVO.getStatus() == TreatmentInfoApplyStatus.APPLY_NOT_PASSED.getVal())
-                .collect(Collectors.toList()));
-        results.addAll(unExpiredAndUnFinishedAppointmentInfoVOS.stream()
-                .filter(treatmentVO -> treatmentVO.getStatus() == TreatmentInfoApplyStatus.WAIT_DOCTOR_APPLY.getVal())
-                .collect(Collectors.toList()));
-
-        List<TreatmentVO> expiredAndFinishedTreatmentVOS = remoteTreatmentMapper.selectExpiredAndFinishedByUserId(userId);
-        results.addAll(expiredAndFinishedTreatmentVOS.stream()
-                .filter(treatmentVO -> treatmentVO.getStatus() == TreatmentInfoApplyStatus.EXPIRED.getVal())
-                .collect(Collectors.toList()));
-        results.addAll(expiredAndFinishedTreatmentVOS.stream().filter(treatmentVO -> treatmentVO.getStatus() == TreatmentInfoApplyStatus.FINISHED.getVal())
-                .collect(Collectors.toList()));
-
+        List<TreatmentVO> treatmentVOS = remoteTreatmentMapper.selectTreatmentByUserId(userId, new Page<>(page, size));
+        Iterator<TreatmentVO> iterator = treatmentVOS.iterator();
+        while (iterator.hasNext()) {
+            TreatmentVO treatmentVO = iterator.next();
+            if (treatmentVO.getStatus() == APPLY_PASSED.getVal() && isCanStartLive(getTreatmentTime(treatmentVO.getDate(), treatmentVO.getStartTime()))) {
+                treatmentVO.setStart(true);
+                results.add(treatmentVO);
+                iterator.remove();
+            }
+        }
+        results.addAll(treatmentVOS);
         results.forEach(treatmentVO -> {
             treatmentVO.setTreatmentTime(getTreatmentTime(treatmentVO.getDate(), treatmentVO.getStartTime()));
             handleDate(treatmentVO);
         });
-        cacheService.delete(RedisCacheKey.USER_TREATMENT_STATUS_CNT_KEY + userId);
+        if (page == 1) {
+            cacheService.delete(RedisCacheKey.USER_TREATMENT_STATUS_CNT_KEY + userId);
+        }
         return results;
     }
